+++ framework/db-local/src/main/java/is/codion/framework/db/local/DefaultLocalEntityConnection.java
@@ -796,6 +796,7 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection, Metho
 		List<Key> insertedKeys = new ArrayList<>(entities.size());
 		List<Object> statementValues = new ArrayList<>();
 		List<ColumnDefinition<?>> statementColumns = new ArrayList<>();
+		Map<EntityType, Set<Column<?>>> lazyInsertedColumns = new HashMap<>();
 		String insertQuery = null;
 		synchronized (database) {
 			try {
@@ -806,7 +807,8 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection, Metho
 						generateBeforeInsert(entity, column);
 					}
 					populateColumnsAndValues(entity, insertableColumns(entityDefinition),
-									statementColumns, statementValues, column -> entity.contains(column.attribute()));
+									statementColumns, statementValues, column -> entity.contains(column.attribute()),
+									lazyInsertedColumns.computeIfAbsent(entity.type(), k -> new HashSet<>()));
 					if (generatedColumns.isEmpty() && statementColumns.isEmpty()) {
 						throw new SQLException("Unable to insert entity " + entity.type() + ", no values to insert");
 					}
@@ -823,8 +825,10 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection, Metho
 					statementValues.clear();
 				}
 				if (insertedEntities != null) {
-					for (List<Key> entityTypeKeys : groupByType(insertedKeys).values()) {
-						insertedEntities.addAll(query(where(keys(entityTypeKeys)).build(), 0));
+					for (Map.Entry<EntityType, List<Key>> entry : groupByType(insertedKeys).entrySet()) {
+						insertedEntities.addAll(query(where(keys(entry.getValue()))
+										.include(lazyInsertedColumns.get(entry.getKey()))
+										.build(), 0));
 					}
 				}
 				commitIfTransactionIsNotOpen();
@@ -871,9 +875,10 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection, Metho
 					List<ColumnDefinition<?>> updatableColumns = updatableColumns(entityDefinition);
 
 					List<Entity> entitiesToUpdate = entityTypeEntities.getValue();
+					Set<Column<?>> lazyUpdatedColumns = new HashSet<>();
 					for (Entity entity : entitiesToUpdate) {
 						populateColumnsAndValues(entity, updatableColumns, statementColumns, statementValues,
-										columnDefinition -> entity.modified(columnDefinition.attribute()));
+										columnDefinition -> entity.modified(columnDefinition.attribute()), lazyUpdatedColumns);
 						if (statementColumns.isEmpty()) {
 							throw new UpdateException("Unable to update entity " + entity.type() + ", no modified values found");
 						}
@@ -893,7 +898,9 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection, Metho
 						statementValues.clear();
 					}
 					if (updatedEntities != null) {
-						List<Entity> selected = query(where(keys(primaryKeys(entitiesToUpdate))).build(), 0);//bypass caching
+						List<Entity> selected = query(where(keys(primaryKeys(entitiesToUpdate)))
+										.include(lazyUpdatedColumns)
+										.build(), 0);//bypass caching
 						if (selected.size() != entitiesToUpdate.size()) {
 							throw new UpdateException(entitiesToUpdate.size() + " updated rows expected, query returned " +
 											selected.size() + ", entityType: " + entityTypeEntities.getKey());
@@ -1603,17 +1610,22 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection, Metho
 	 * @param statementColumns the list to populate with the columns to use in the statement
 	 * @param statementValues the list to populate with the values to be used in the statement
 	 * @param includeIf the Predicate to apply when checking to see if the column should be included
+	 * @param lazy populated with the lazy columns included in the operation
 	 */
 	private static void populateColumnsAndValues(Entity entity,
 																							 List<ColumnDefinition<?>> columnDefinitions,
 																							 List<ColumnDefinition<?>> statementColumns,
 																							 List<@Nullable Object> statementValues,
-																							 Predicate<ColumnDefinition<?>> includeIf) {
+																							 Predicate<ColumnDefinition<?>> includeIf,
+																							 Set<Column<?>> lazy) {
 		for (int i = 0; i < columnDefinitions.size(); i++) {
 			ColumnDefinition<?> columnDefinition = columnDefinitions.get(i);
 			if (includeIf.test(columnDefinition)) {
 				statementColumns.add(columnDefinition);
 				statementValues.add(entity.get(columnDefinition.attribute()));
+				if (!columnDefinition.selected()) {
+					lazy.add(columnDefinition.attribute());
+				}
 			}
 		}
 	}